# perf: Optimize performance of `LocalizationProvider.Translate<T>` method